Makefile.legacy: Use modules.sample as the default /etc/dahdi/module file.
This was moved to only be installed as /etc/dahdi/modules.sample file as part
of the changes to use autoconf. The problem is that there are several workflows
where new systems are installed with the make config target without knowing the
module.sample needs to be renamed.

Restore the old behavior until more code can be put in to help users know what
needs to be changed, etc, as part of the automake changes since the release
notes did not mention anything about these changes being necessary.

Signed-off-by: Shaun Ruffell <sruffell@digium.com>